Changed overall state change monitoring
Rotary encoder+switch are not monitor when state is not Configuration(_).
When entering configuration mode, start a polling task for both rotary encoder
and switch (no more interrupt here).

Toggle switch is still monitored using interrupt: a debounce task is started
when an IT is raised and the task is not already running/queued.